Rename `read_paths.py` to `read_prep.py`
